requirements-dev.txt
astroid==2.3.3
bcrypt==3.1.7
cffi==1.14.0
cryptography==2.9.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
paramiko==2.7.1
pycparser==2.20
pylint==2.4.4
PyNaCl==1.3.0
PyYAML==5.3.1
six==1.14.0
sshtunnel==0.1.5
wrapt==1.11.2
zope.interface==5.1.0